+++ tracking.py
@@ -15,6 +15,21 @@ from exceptions import error_popup, warning_popup
 from enums import *
 
 def scale_frame(frame, scale_factor=0.9):
+    """Sometimes resolution of video is larger than resolution of monitor this software is running on.
+    This function scales the resolution of the frame so the entire frame can be seen for selecting markers or just general analysis.
+    Scaling preseves aspect ratio, so it determines which dimension is most cut off from viewing (width or height),
+    and determines caling ratio for the other dimension based on that.
+
+    Distances of objects being tracked are scaled back up when recording data, so movement of tracked objects are recorded in the original resolution.
+
+    Args:
+        frame (numpy.ndarray): indv frame of video being tracked that will be scaled and returned
+        scale_factor (float, optional): fraction of monitor resolution to scale image. Defaults to 0.9.
+
+    Returns:
+        scaled_frame (numpy.ndarray): scaled version of frame that was passed in
+        min_scale_factor (float): Determing scale factor, used to scale values back up before recording data.
+    """    
     monitor = screeninfo.get_monitors()[0] # get primary monitor resolution
 
     # get indv scale factors for width and height
@@ -69,6 +84,8 @@ def select_markers(cap, bbox_size, frame_start):
 
     Args:
         cap (cv2.VideoCapture): loaded video file for selecting markers of
+        bbox_size (int): size of bounding box that is spec'd by user in ui
+        frame_start (int): starting frame of video indicated in the frame selection tool
 
     Returns:
         mouse_params['marker_positions'] (list): selected marker positions
@@ -108,20 +125,37 @@ def track_markers(
         cap,
         bbox_size,
         tracker_choice,
+        frame_record_interval,
         frame_interval,
         time_units,
         file_mode,
         video_file_name,
         data_label):
-    """main tracking loop of markers selected using marker selections from select_markers()
-    saves distances of each mark each frame update to 'output/Tracking_Output.csv'
+    """Implements the main tracking loop for markers on a video.
+    It tracks markers from a user-defined start to end frame and records the data.
+
+    Functionality:
+        1. Initializes trackers based on user-selected markers and tracking algorithm.
+        2. Scales the first frame and initializes bounding boxes for trackers.
+        3. Enters a loop to track markers across specified frames, scaling each frame as processed.
+        4. Records tracking data for each frame where markers are successfully tracked.
+        5. Saves or appends the tracking data to a CSV file depending on the file mode.
+        6. Handles errors by displaying warning messages and terminating tracking if necessary.
 
     Args:
-        marker_positions (list): list of marker positions from user selections
-        first_frame (np.Array(np.uint8)): image of first frame of loaded video
-        cap (cv2.VideoCapture): loaded video
-        bbox_size(int): length of 1 edge of the bounding box that will be the size of the tracker
-        tracker_choice(str): 'KCF' or 'CSRT' determines choice of tracking algorithm used (see README.md for details)
+        marker_positions (list of tuples): Initial positions of markers selected by the user.
+        first_frame (np.array): Image of the first frame of the video where markers were selected.
+        frame_start (int): The frame number to start tracking.
+        frame_end (int): The frame number to end tracking.
+        cap (cv2.VideoCapture): Video capture object loaded with the video file.
+        bbox_size (int): Size of the bounding box for each tracker.
+        tracker_choice (TrackerChoice): Enum specifying the tracking algorithm to use (KCF or CSRT).
+        frame_record_interval (int): Interval at which frames are processed and data is recorded.
+        frame_interval (float): Real-world time interval between frames, used in time calculations for data recording.
+        time_units (str): Units for time (e.g., 'seconds', 'minutes') used in the output data.
+        file_mode (FileMode): Specifies whether to overwrite or append data in the output file.
+        video_file_name (str): Name of the video file being processed.
+        data_label (str): Unique identifier for the tracking session, used in data labeling.
     """    
     # create trackers
     trackers = []
@@ -148,7 +182,9 @@ def track_markers(
     # tracking loop
     while True:
         ret, frame = cap.read()
-        frame_num += 1
+        frame_num += frame_record_interval
+        cap.set(cv2.CAP_PROP_POS_FRAMES, frame_num)
+
         if not ret:
             break  # break when frame read unsuccessful (end of video or error)
 
@@ -182,7 +218,7 @@ def track_markers(
 
         cv2.imshow("Tracking...", scaled_frame)  # show updated frame tracking
 
-        if cv2.waitKey(1) == 27 or frame_num == frame_end - frame_start:  # cut tracking loop short if ESC hit
+        if cv2.waitKey(1) == 27 or frame_num >= frame_end - frame_start:  # cut tracking loop short if ESC hit
             break
     
     if file_mode == FileMode.OVERWRITE:
@@ -223,6 +259,7 @@ def necking_point(
         percent_crop_left,
         percent_crop_right,
         binarize_intensity_thresh,
+        frame_record_interval,
         frame_interval,
         time_units,
         file_mode,
@@ -231,18 +268,30 @@ def necking_point(
     ):
 
     """necking point detection loop
-    necking point defined as the most shortest vertical line between two horizontal edges
-    frames are preprocessed and then edges are detected, top and bottom most edges are singled out
-    each frame the vertical distance between these edges are recorded at every x coordinate and the shortest is highlighted in red
-    in cases where there are multiple min distance values, the median x location is chosen
-    recorded data saved to "output/Necking_Point_Output.csv"
+    Detects and records the necking point in each frame of a video. The necking point is defined as the shortest vertical distance 
+    between two horizontal edges within a specified region of interest. The function preprocesses the frames, performs edge detection, 
+    and identifies the top and bottom edges to calculate the necking point, highlighting it in red on the visual output.
+
+    Functionality:
+        1. Processes each frame based on the specified interval, applying grayscale conversion, binarization, and edge detection.
+        2. Identifies and calculates the shortest vertical distance between detected edges.
+        3. Records the x-coordinate and the vertical distance of the necking point in each frame.
+        4. Outputs visual representation with necking points highlighted.
+        5. Saves the recorded data to a CSV file, either overwriting or appending, based on the file mode.
 
     Args:
-        cap (cv2.VideoCapture): loaded video file for selecting markers of
-        percent_crop_left (float, optional): percentage of pixels to remove from consideration from the left side of the frame of necking pt detection. Defaults to 0..
-        percent_crop_right (_type_, optional): percentage of pixels to remove from consideration from the right side of the frame of necking pt detection. Defaults to 0..
-        binarize_intensity_thresh (int, optional): threshold pixel intensity value for frame binarization. Defaults to 120.
-        x_interval (int, optional): interval of horizontal pixels to draw vertical blue lines for visualization purposes. Defaults to 50.
+        cap (cv2.VideoCapture): Video capture object loaded with the video.
+        frame_start (int): Frame number to start processing.
+        frame_end (int): Frame number to end processing.
+        percent_crop_left (float): Percentage of the frame's left side to exclude from processing.
+        percent_crop_right (float): Percentage of the frame's right side to exclude from processing.
+        binarize_intensity_thresh (int): Threshold for binarization of the frame to facilitate edge detection.
+        frame_record_interval (int): Interval at which frames are processed.
+        frame_interval (float): Real-world time interval between frames, used in time calculations.
+        time_units (str): Units of time (e.g., 'seconds', 'minutes') for the output data.
+        file_mode (FileMode): Specifies whether to overwrite or append data in the output file.
+        video_file_name (str): Name of the video file being processed.
+        data_label (str): Unique identifier for the data session.
     """    
     x_interval = 50 # interval for how many blue line visuals to display
     frame_num = 0
@@ -252,8 +301,9 @@ def necking_point(
 
     while True:  # read frame by frame until the end of the video
         ret, frame = cap.read()
+        frame_num += frame_record_interval
         cap.set(cv2.CAP_PROP_POS_FRAMES, frame_start + frame_num)
-        frame_num += 1
+
         if not ret:
             break
 
@@ -319,7 +369,7 @@ def necking_point(
         cv2.line(frame_draw, (x_samples[necking_pt_ind], y_line_values[necking_pt_ind][0]), (x_samples[necking_pt_ind], y_line_values[necking_pt_ind][1]), (0, 0, 255), 2)     
 
         cv2.imshow('Necking Point Visualization', frame_draw)
-        if cv2.waitKey(1) == 27 or frame_end == frame_num+frame_start:
+        if cv2.waitKey(1) == 27 or frame_end <= frame_num + frame_start:
             break
 
     if file_mode == FileMode.OVERWRITE:
@@ -365,6 +415,33 @@ def noise_reduction(frame):
 
 
 def improve_binarization(frame):    
+    """
+    Enhances the binarization of a grayscale image using various image processing techniques. This function applies
+    CLAHE for contrast enhancement, background subtraction to highlight foreground objects, morphological operations
+    to refine the image, and edge detection to further define object boundaries.
+
+    Steps:
+        1. Apply Contrast Limited Adaptive Histogram Equalization (CLAHE) to boost the contrast of the image.
+        2. Perform background subtraction using a median blur to isolate foreground features.
+        3. Apply morphological closing to close small holes within the foreground objects.
+        4. Detect edges using the Canny algorithm, and dilate these edges to enhance their visibility.
+        5. Optionally adjust the edge thickness with additional morphological operations like dilation or erosion
+           depending on specific requirements (commented out in the code but can be adjusted as needed).
+
+    Note:
+        - This function is designed to work with grayscale images and expects a single-channel input.
+        - Adjustments to parameters like CLAHE limits, kernel sizes for morphological operations, and Canny thresholds
+          may be necessary depending on the specific characteristics of the input image.
+
+
+    Args:
+        frame (np.array): A single-channel (grayscale) image on which to perform binarization improvement.
+
+    Returns:
+        np.array: The processed image with enhanced binarization and clearer object definitions.
+
+    """
+
     # Apply CLAHE (Contrast Limited Adaptive Histogram Equalization)
     # boosts contrast
     clahe = cv2.createCLAHE(clipLimit=2.0, tileGridSize=(12, 12))
@@ -400,6 +477,7 @@ def track_area(
         bbox_size,
         frame_start,
         frame_end,
+        frame_record_interval,
         frame_interval,
         time_units,
         distance_from_marker_thresh,
@@ -407,6 +485,40 @@ def track_area(
         video_file_name,
         data_label
     ):
+
+    """
+    Tracks the surface area of an object in a video, given initial marker positions. This function initializes trackers on the object,
+    processes each frame to apply binarization and contour detection, and records the largest contour's area that is close to the marker.
+
+    Functionality:
+        1. Initializes trackers using CSRT algorithm at the marker positions in the scaled first frame.
+        2. Processes each frame: scales, converts to grayscale, and applies binarization improvements.
+        3. Applies adaptive thresholding to the binary frame to enhance contour detection.
+        4. Identifies contours and selects the optimal one based on area size and proximity to the initial marker.
+        5. Draws the contour and marker on the frame for visualization.
+        6. Records the position and area of the selected contour.
+        7. Saves or appends the tracking data to a CSV file based on the specified file mode.
+        8. Displays the tracking in real-time and provides an option to exit by pressing ESC.
+
+    Note:
+        - Ensure the video file is properly loaded and parameters are correctly set for accurate tracking.
+        - Binarization and contour detection settings might need adjustments based on different video characteristics.
+
+    Args:
+        cap (cv2.VideoCapture): Video capture object loaded with the video.
+        marker_positions (list of tuples): Initial positions of markers.
+        first_frame (np.array): First frame of the video for initializing the tracker.
+        bbox_size (int): Size of the bounding box for the tracker.
+        frame_start (int): Starting frame number for tracking.
+        frame_end (int): Ending frame number for tracking.
+        frame_record_interval (int): Number of frames to skip between recordings.
+        frame_interval (float): Real-world time interval between frames, used in time calculations.
+        time_units (str): Units of time (e.g., 'seconds', 'minutes') for the output data.
+        distance_from_marker_thresh (int): Threshold for determining proximity of the contour to the marker.
+        file_mode (FileMode): Specifies whether to overwrite or append data in the output file.
+        video_file_name (str): Name of the video file being processed.
+        data_label (str): Unique identifier for the data session.
+        """
 
     cap.set(cv2.CAP_PROP_POS_FRAMES, frame_start)
     frame_num = frame_start
@@ -428,6 +540,9 @@ def track_area(
 
     while frame_num < frame_end:
         ret, frame = cap.read()
+        frame_num += frame_record_interval
+        cap.set(cv2.CAP_PROP_POS_FRAMES, frame_num)
+
         if not ret:
             break